raspsecurity/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.11
numpy==2.4.2
pillow==12.1.0
requests==2.32.5
urllib3==2.6.3
